+++ packages/core/CHANGELOG.md
@@ -1,3 +1,10 @@
+# @qiwi/pijma-core [1.134.0](https://github.com/qiwi/pijma/compare/@qiwi/pijma-core@1.133.0...@qiwi/pijma-core@1.134.0) (2021-11-01)
+
+
+### Features
+
+* no render props ([#445](https://github.com/qiwi/pijma/issues/445)) ([5da89bd](https://github.com/qiwi/pijma/commit/5da89bd1154a6eeaef78d728e40b1ce2162b2f6a))
+
 # @qiwi/pijma-core [1.133.0](https://github.com/qiwi/pijma/compare/@qiwi/pijma-core@1.132.1...@qiwi/pijma-core@1.133.0) (2021-10-19)